makin good progress on refactor with data-utils
