+++ src/services/chart-blueprints.ts
@@ -1,0 +1,6 @@
+export type ChartBlueprint = {
+  readonly label: string
+  readonly component: any
+}
+
+// TODO
