+++ src/views/ActionCreateView.vue
@@ -1,34 +1,18 @@
 <script setup lang="ts">
 import { Icon } from '@/types/icons'
-import { DatabaseField } from '@/types/database'
-import {
-  getFields,
-  getDatabaseTypeColumnProps,
-  getLabelSingular,
-  getTypeFromSlug,
-} from '@/services/DatabaseUtils'
-import { useRoute } from 'vue-router'
+import { DatabaseField, DatabaseType } from '@/types/database'
 import type { DatabaseRecord } from '@/types/models'
+import { getFieldBlueprints, getFields } from '@/services/data-utils'
 import ResponsivePage from '@/components/ResponsivePage.vue'
 import useActions from '@/composables/useActions'
+import useRouteParams from '@/composables/useRouteParams'
 import useActionRecordStore from '@/stores/action-record'
-import ActionInputId from '@/components/ActionInputId.vue'
-import ActionInputCreatedTimestamp from '@/components/ActionInputCreatedTimestamp.vue'
-import ActionInputName from '@/components/ActionInputName.vue'
-import ActionInputText from '@/components/ActionInputText.vue'
-import ActionInputParentId from '@/components/ActionInputParentId.vue'
-import ActionInputToggle from '@/components/ActionInputToggle.vue'
-import ActionInputNumber from '@/components/ActionInputNumber.vue'
 
-const route = useRoute()
+const { routeDatabaseType, routeParentId } = useRouteParams()
 const { onCreateRecord } = useActions()
 const actionRecordStore = useActionRecordStore()
 
-const routeDatabaseType = getTypeFromSlug(route?.params?.databaseTypeSlug as string)
-const routeParentId = route?.params?.[DatabaseField.PARENT_ID] as string
-const bannerTitle = `Create ${getLabelSingular(routeDatabaseType)}`
-
-const columnProps = getDatabaseTypeColumnProps(routeDatabaseType)
+const fieldBlueprints = getFieldBlueprints(routeDatabaseType as DatabaseType)
 
 // TODO - Must do this before Create and Edit actions
 actionRecordStore.$reset()
@@ -36,7 +20,7 @@ actionRecordStore.temp[DatabaseField.TYPE] = routeDatabaseType
 
 // TODO
 async function onCreate(record: DatabaseRecord) {
-  const fields = getFields(routeDatabaseType)
+  const fields = getFields(routeDatabaseType as DatabaseType)
   const realRecord = {} as DatabaseRecord
   // removing null fields from store record before creation
   fields.forEach((field) => {
@@ -47,9 +31,12 @@ async function onCreate(record: DatabaseRecord) {
 </script>
 
 <template>
-  <ResponsivePage :banner-icon="Icon.CREATE" :banner-title="bannerTitle">
-    <div v-for="(item, i) in columnProps" :key="i" class="q-mb-md">
-      <!-- Do NOT print the hiddenId field -->
+  <ResponsivePage :banner-icon="Icon.CREATE" banner-title="Create">
+    <div v-for="(fieldBP, i) in fieldBlueprints" :key="i" class="q-mb-md">
+      <component :is="fieldBP.component" />
+    </div>
+
+    <!-- <div v-for="(item, i) in columnProps" :key="i" class="q-mb-md">
       <div v-if="item.name !== 'hiddenId'">
         <ActionInputId v-if="item.name === DatabaseField.ID" />
         <ActionInputCreatedTimestamp
@@ -74,7 +61,7 @@ async function onCreate(record: DatabaseRecord) {
         />
         <ActionInputNumber v-if="item.name === DatabaseField.NUMBER" />
       </div>
-    </div>
+    </div> -->
 
     <QBtn
       label="Create"
