+++ src/views/ActionEditView.vue
@@ -1,40 +1,24 @@
 <script setup lang="ts">
 import { Icon } from '@/types/icons'
 import { DatabaseField, DatabaseType } from '@/types/database'
-import {
-  getFields,
-  getDatabaseTypeColumnProps,
-  getLabelSingular,
-  getTypeFromSlug,
-} from '@/services/DatabaseUtils'
-import { useRoute } from 'vue-router'
 import type { DatabaseRecord } from '@/types/models'
 import { onMounted, type Ref, ref } from 'vue'
 import type { Optional } from '@/types/misc'
 import useLogger from '@/composables/useLogger'
 import useDatabase from '@/composables/useDatabase'
 import useActions from '@/composables/useActions'
+import useRouteParams from '@/composables/useRouteParams'
 import useActionRecordStore from '@/stores/action-record'
 import ResponsivePage from '@/components/ResponsivePage.vue'
-import ActionInputId from '@/components/ActionInputId.vue'
-import ActionInputCreatedTimestamp from '@/components/ActionInputCreatedTimestamp.vue'
-import ActionInputName from '@/components/ActionInputName.vue'
-import ActionInputText from '@/components/ActionInputText.vue'
-import ActionInputParentId from '@/components/ActionInputParentId.vue'
-import ActionInputToggle from '@/components/ActionInputToggle.vue'
-import ActionInputNumber from '@/components/ActionInputNumber.vue'
+import { getFieldBlueprints, getFields } from '@/services/data-utils'
 
-const route = useRoute()
+const { routeDatabaseType, routeId } = useRouteParams()
 const { log } = useLogger()
 const { onUpdateRecord } = useActions()
 const actionRecordStore = useActionRecordStore()
 const { getRecord } = useDatabase()
 
-const routeDatabaseType = getTypeFromSlug(route?.params?.databaseTypeSlug as string)
-const routeId = route?.params?.[DatabaseField.ID] as string
-const bannerTitle = `Edit ${getLabelSingular(routeDatabaseType)}`
-
-const columnProps = getDatabaseTypeColumnProps(routeDatabaseType)
+const fieldBlueprints = getFieldBlueprints(routeDatabaseType as DatabaseType)
 
 // TODO - Must do this before Create and Edit actions
 actionRecordStore.$reset()
@@ -44,12 +28,12 @@ const oldRecord: Ref<Optional<DatabaseRecord>> = ref(null)
 
 onMounted(async () => {
   // TODO
-  oldRecord.value = await getRecord(routeDatabaseType, routeId)
+  oldRecord.value = await getRecord(routeDatabaseType as DatabaseType, routeId)
 })
 
 // TODO
 async function onUpdate(type: DatabaseType, originalId: string, record: DatabaseRecord) {
-  const fields = getFields(routeDatabaseType)
+  const fields = getFields(routeDatabaseType as DatabaseType)
   const realRecord = {} as DatabaseRecord
   // removing null fields from store record before creation
   fields.forEach((field) => {
@@ -60,9 +44,12 @@ async function onUpdate(type: DatabaseType, originalId: string, record: Database
 </script>
 
 <template>
-  <ResponsivePage :banner-icon="Icon.CREATE" :banner-title="bannerTitle">
-    <div v-for="(item, i) in columnProps" :key="i" class="q-mb-md">
-      <!-- Do NOT print the hiddenId field -->
+  <ResponsivePage :banner-icon="Icon.EDIT" banner-title="Edit">
+    <div v-for="(fieldBP, i) in fieldBlueprints" :key="i" class="q-mb-md">
+      <component :is="fieldBP.component" />
+    </div>
+
+    <!-- <div v-for="(item, i) in columnProps" :key="i" class="q-mb-md">
       <div v-if="item.name !== 'hiddenId'">
         <ActionInputId v-if="item.name === DatabaseField.ID" />
         <ActionInputCreatedTimestamp v-if="item.name === DatabaseField.CREATED_TIMESTAMP" />
@@ -82,13 +69,13 @@ async function onUpdate(type: DatabaseType, originalId: string, record: Database
         />
         <ActionInputNumber v-if="item.name === DatabaseField.NUMBER" />
       </div>
-    </div>
+    </div> -->
 
     <QBtn
       label="Update"
       color="positive"
       :icon="Icon.SAVE"
-      @click="onUpdate(routeDatabaseType, routeId, { ...actionRecordStore.temp })"
+      @click="onUpdate(routeDatabaseType as DatabaseType, routeId, { ...actionRecordStore.temp })"
     />
   </ResponsivePage>
 </template>
